requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
pydantic==2.4.2
supabase==1.2.0
python-multipart==0.0.6
geopy==2.4.0
folium==0.14.0
pytest==7.4.3